User Interface aesthetics and follows input commands

+++ NIM_Patient_Sim/GpioControl/spicontrol.h
@@ -20,11 +20,10 @@ public:
     int spi_Write_DAC(uint8_t reg, uint16_t write_data);
     int spi_Read_ADC(uint8_t reg, uint8_t *data);
     int spi_Write_ADC(uint8_t reg, uint8_t data);
-    int DAC_output(uint8_t output, float_t gain, uint16_t delay_time_ms, uint16_t amplitude_uV);
+    int DAC_output(uint8_t output, float_t gain, uint16_t delay_time_ms, uint32_t amplitude_uV);
 
 
 public:
-    //Ui::SpiControl *ui;
 
 signals:
